280 update and add validation rules for v3 pipelines and actions (#304)
* refactor: In the process of introducing the expr validation in validator v3

* feat: Added symbol validation for v3 validator using the expression evaluator

* style: Changes to resolve lint warnings

* refactor: Added hash maps to the inputs and env during validator construction

* refactor: Changes to apply proper validation rules

* refactor: Added keyword handling in action eval

* refactor: Changes in order to handle ignore outputs from validation

* refactor: Moved writtable context into the validator module

* refactor: Minor rename

* feat: Added additional lints

* feat: Added extra lints

* ci: Updated rust to 1.95

+++ crates/bld_runner/src/expr/v3/parser.rs
@@ -234,7 +234,7 @@ mod tests {
                 for (i, object_part) in object_parts.into_iter().enumerate() {
                     assert_eq!(object_part.as_rule(), Rule::ObjectPart);
                     assert_eq!(object_part.as_span().as_str(), spans[i]);
-                    assert_eq!(object_part.into_inner().count(), 0 as usize);
+                    assert_eq!(object_part.into_inner().count(), 0_usize);
                 }
             }
         }
@@ -292,7 +292,7 @@ mod tests {
                             for (i, object_part) in object_parts.into_iter().enumerate() {
                                 assert_eq!(object_part.as_rule(), Rule::ObjectPart);
                                 assert_eq!(object_part.as_span().as_str(), symbol[i + 1]);
-                                assert_eq!(object_part.into_inner().count(), 0 as usize);
+                                assert_eq!(object_part.into_inner().count(), 0_usize);
                             }
                         }
